requirements.txt
email-validator>=2.2.0
flask-mail>=0.10.0
flask>=3.1.1
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
psycopg2-binary>=2.9.10
python-dotenv>=1.0.0